Improved error handling

+++ src/main/java/mx/uaemex/fi/linc13/saturn/engine/FRGraphRenderer.java
@@ -2,6 +2,7 @@ package mx.uaemex.fi.linc13.saturn.engine;
 
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
+
 import mx.uaemex.fi.linc13.saturn.data.Graph;
 import mx.uaemex.fi.linc13.saturn.data.Graph.Edge;
 import mx.uaemex.fi.linc13.saturn.data.Graph.Vertex;
@@ -11,11 +12,11 @@ public class FRGraphRenderer {
     private static Random random = ThreadLocalRandom.current();
 
     public static void computeVertexPositions(Graph graph, int w, int h){
-        double temp = w /10;
-        double k = (double) Math.sqrt((w*h)/graph.vertexes.size());
+        double temp = w/10;
+        double k = Math.sqrt((w*h)/graph.vertexes.size());
         graph.vertexes.forEach(v -> v.updatePos(random.nextInt(w), random.nextInt(h)));
 
-        for(double i = 0; i < 50; i++){
+        for(int i = 0; i < 50; i++){
             for (Vertex v : graph.vertexes){
                 v.updateDisp(0, 0);
 
@@ -23,7 +24,7 @@ public class FRGraphRenderer {
                     if(!v.equals(u)){
                         double d1 = v.getPosX() - u.getPosX();
                         double d2 = v.getPosY() - u.getPosY();
-                        double dlen = (double) Math.sqrt(d1*d1 +d2*d2);
+                        double dlen = Math.sqrt(d1*d1 + d2*d2);
 
                         v.updateDisp(v.getDispX() + (d1/dlen)*(k*k/dlen), v.getDispY() + (d2/dlen)*(k*k/dlen));
                     }
@@ -33,7 +34,7 @@ public class FRGraphRenderer {
             for(Edge e : graph.edges){
                 double d1 = e.A.getPosX() - e.B.getPosX();
                 double d2 = e.A.getPosY() - e.B.getPosY();
-                double dlen = (double) Math.sqrt(d1*d1 +d2*d2);
+                double dlen = Math.sqrt(d1*d1 + d2*d2);
 
                 e.A.updateDisp(e.A.getDispX() - (d1/dlen)*(dlen*dlen/k), e.A.getDispY() - (d2/dlen)*(dlen*dlen/k));
                 e.B.updateDisp(e.B.getDispX() + (d1/dlen)*(dlen*dlen/k), e.B.getDispY() + (d2/dlen)*(dlen*dlen/k));
@@ -42,7 +43,7 @@ public class FRGraphRenderer {
             for(Vertex v : graph.vertexes){
                 double d1 = v.getDispX();
                 double d2 = v.getDispY();
-                double dlen = (double) Math.sqrt(d1*d1 + d2*d2);
+                double dlen = Math.sqrt(d1*d1 + d2*d2);
                 double min = Math.min(dlen, temp);
 
                 if (dlen > 0) {
@@ -52,10 +53,10 @@ public class FRGraphRenderer {
                     );
                 }
 
-                v.updatePos((double)Math.min(w, Math.max(0, v.getPosX())), (double)Math.min(h, Math.max(0, v.getPosY())));
+                v.updatePos(Math.min(w, Math.max(0, v.getPosX())), Math.min(h, Math.max(0, v.getPosY())));
             }
 
-            temp = (double) Math.max(temp * (1 - 0.065), 1);
+            temp = Math.max(temp * (1 - 0.065), 1);
         }
     }
